Rename AbstractPage/AbstractTest to BasePage/BaseTest
