+++ src/main/java/org/diplom3/pages/AccountPage.java
@@ -5,7 +5,7 @@ import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
-public class AccountPage extends AbstractPage {
+public class AccountPage extends BasePage {
 
     public final WebDriver driver;
