+++ src/main/java/org/diplom3/pages/ConstructorPage.java
@@ -5,7 +5,7 @@ import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
-public class ConstructorPage extends AbstractPage {
+public class ConstructorPage extends BasePage {
     public final WebDriver driver;
 
     public ConstructorPage(WebDriver driver){
