+++ src/main/java/org/diplom3/pages/LoginPage.java
@@ -5,7 +5,7 @@ import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
-public class LoginPage extends AbstractPage {
+public class LoginPage extends BasePage {
 
     public final WebDriver driver;
