+++ src/main/java/org/diplom3/pages/RegistrationPage.java
@@ -7,7 +7,7 @@ import org.openqa.selenium.support.ui.WebDriverWait;
 
 import java.util.Random;
 
-public class RegistrationPage extends AbstractPage {
+public class RegistrationPage extends BasePage {
     public final WebDriver driver;
 
     public RegistrationPage(WebDriver driver){
